+++ crates/cli/Cargo.toml
@@ -37,7 +37,7 @@ proto_installer = { version = "0.2.0", path = "../installer" }
 proto_pdk_api = { version = "0.14.2", path = "../pdk-api" }
 proto_shim = { version = "0.2.6", path = "../shim" }
 system_env = { version = "0.2.0", path = "../system-env" }
-anyhow = { workspace = true }
+anyhow = "1.0.79"                                                      # For shims
 chrono = "0.4.31"
 clap = { workspace = true, features = ["derive", "env"] }
 clap_complete = { workspace = true }
