+++ crates/pdk-api/Cargo.toml
@@ -11,7 +11,6 @@ repository = "https://github.com/moonrepo/proto"
 system_env = { version = "0.2.0", path = "../system-env" }
 version_spec = { version = "0.2.0", path = "../version-spec" }
 warpgate_api = { version = "0.2.0", path = "../warpgate-api" }
-anyhow = { workspace = true }
 semver = { workspace = true, features = ["serde"] }
 serde = { workspace = true }
 serde_json = { workspace = true }
