+++ crates/pdk/Cargo.toml
@@ -9,6 +9,8 @@ repository = "https://github.com/moonrepo/proto"
 
 [dependencies]
 proto_pdk_api = { version = "0.14.2", path = "../pdk-api" }
-anyhow = { workspace = true }
+warpgate_api = { version = "0.2.0", path = "../warpgate-api", features = [
+	"pdk",
+] }
 extism-pdk = { workspace = true }
 serde = { workspace = true }
