+++ crates/pdk/src/lib.rs
@@ -3,3 +3,4 @@ mod macros;
 
 pub use helpers::*;
 pub use proto_pdk_api::*;
+pub use warpgate_api::pdk::*;
